requirements.txt
algoliasearch==3.0.0
algoliasearch-django==3.0.0
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
djangorestframework==3.14.0
idna==3.6
numpy==1.26.3
pandas==2.2.0
pillow==11.0.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
